rename components (fix comments)
