+++ kiwiapp-public/src/components/overview/index.ts
@@ -2,4 +2,4 @@ export { OverviewCard } from "./OverviewCard";
 export { SelectedBar } from "./SelectedBar";
 export { ChartCard } from "./ChartCard";
 export { BarChart } from "./BarChart";
-export { DoughnutChart } from "./DoughnutChart";
+export { DonutChart } from "./DonutChart";
