+++ kiwiapp-public/src/components/select/SelectDropdownMultiple.tsx
@@ -77,7 +77,7 @@ SelectDropdownMultipleProps<T>) => {
           <ComboboxOption
             key={index}
             value={option.value}
-            as='div'
+            // as='div'
             onClick={() => toggleSelection(option.value)}
             className={clsx(
               "flex items-center gap-2 cursor-pointer px-3 py-1.5 rounded-lg",
